Update patch set 4
Patch Set 4:

It's not so much about patches to the drivers, but about the configuration of the interactive and scheduler parameters.

I'm not entirely sure about this, but I think that the updated MSM8939 v3.0 parameters are meant to go together with the enabling of the interactive scheduler input feature.

https://us.codeaurora.org/cgit/quic/la/device/qcom/common/tree/rootdir/etc/init.qcom.post_boot.sh?h=LA.BR.1.2.6_rb1.12#n703 (stuff like the different configuration of the scheduler window stats, task migration, and interactive governor)

Patch-set: 4